# hacktoberfest/docs/api.yml
openapi: '3.0.3'
info:
  title: Library management API
  version: '1.0'
servers:
  - url: https://library-management.com
security:
  - BearerAuth: []
paths:
...
  /authorize:
    put:
      tags:
        - Security
      summary: Update user if id provided, create it otherwise
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        200:
          description: The Created or Updated user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
    post:
      tags:
        - Security
      summary: How to get authentication token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: hacktoberfest
                password:
                  type: string
                  example: "*******"
      responses:
        200:
          description: The authentication token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiJ9"
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
...
components: 
  schemas:
    BadRequest:
      description: Bad request exception
      type: object
      properties:
        type:
          type: string
          example: BadRequestException
        message:
          type: string
          example: Bad request
...
    User: 
      description: Description of an user
      type: object
      properties:
        id: 
          type: integer
          example: 1
        name:
          type: string
          example: hacktoberfest
        phoneNumber:
          type: string
          example: +261 32*******
        email: 
          type: string 
          example: hacktoberfest@*****.com
        userName:
          type: string 
          example: Unknown
        Password:
          type: string
          example: "*********"
        Role: 
          type: string 
          example: ADMIN